Renamed the ProfileScreen class to SettingsScreen
